added vimdc cleaned up install.sh cleaned up ifips added rsync cleaned virtual box

+++ install.sh
@@ -6,16 +6,27 @@ HOSTFILE=".`hostname`"
 if [[ -f ~/.bash_profile || -f ~/.bashrc ]]; then
 	mkdir $BACKUP_DIR
 	mv ~/.bash* $BACKUP_DIR
-	mv ~/.inputrc $BACKUP_DIR
+
+	if [[ -f ~/.inputrc ]]; then
+		mv ~/.inputrc $BACKUP_DIR
+	fi
+
+	if [[ -f ~/.vimrc ]]; then
+		mv ~/.vimrc $BACKUP_DIR
+	fi
 
 	BASH_FILES=`ls $SOURCE_DIR/.bash*`
 	for BASH_FILE in $BASH_FILES; do
-		SYMLINK=`basename $BASH_FILE`
-		cp $BASH_FILE ~/$SYMLINK
+		DEST_FILE=`basename $BASH_FILE`
+		cp $BASH_FILE ~/$DEST_FILE
 	done
 
 	if [[ ! -f ~/.rsyncignore_global ]]; then
 		cp $SOURCE_DIR/.rsyncignore_global ~
+	fi
+
+	if [[ ! -f ~/.vimrc ]]; then
+		cp $SOURCE_DIR/.vimrc ~
 	fi
 
 	if [[ ! -f $HOSTFILE ]]; then
